Cleanup - Move OperationImportConfig to separate file - Align indentation in Directory.Packages.props - Change formulation in docs, because we only have a single base class left
